% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.adjacency.R
\name{add.adjacency}
\alias{add.adjacency}
\title{add.adjacency}
\usage{
add.adjacency(g)
}
\arguments{
\item{g}{Must be an igraph object all vertices must have an attributes called "KE_KED" (Key Event Designator), with values of "MIE", "KE", or "AO"  [igraph object]}
}
\value{
identical igraph object as input, with new edge attribute called \code{$adjacency} [igraph object]
}
\description{
Adds \code{$adjacency} edge attribute:
 Algorithm Description
   STEP 1: Potenial non-adjacent KERs identified when distance between KEup and KEdown can be greater than two KEs
   STEP 2: Longest unique paths between "origin" to "terminus" pairs identified.
            NOTE: "Origin" and "terminus" are used instead of MIE/AO to account for incomplete AOPs that do not have the MIE or AO specified yet
   STEP 3: A KER is non-adjacent if its KEs have a possible distance >2 (Step 1) AND DOES NOT occur in any longest unique path (Step 2)
}
