% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.KE_PD.R
\name{add.KE_PD}
\alias{add.KE_PD}
\title{add.KE_PD}
\usage{
add.KE_PD(g)
}
\arguments{
\item{g}{Must be an igraph object all vertices must have an attributes called "KE_KED" (Key Event Designator), with values of "MIE", "KE", or "AO"  [igraph object]}
}
\value{
New igraph object that is identical to input object but now with \code{$KE_PD} vertex attribute [igraph object]
}
\description{
Adds \code{$KE_PD} (KE Path Designator) vertex attribute:
   "origin"    if formally described as MIE, OR degree in = 0
   "terminus"  if formally described as AO, or degree out = 0
   ""          if not origin or terminus
}
\seealso{
Other KE Attribution: \code{\link{add.KE_LAOC}}
}
