% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOPN.edge.connectivity.R
\name{AOPN.edge.connectivity}
\alias{AOPN.edge.connectivity}
\title{AOPN.edge.connectivity}
\usage{
AOPN.edge.connectivity(g, use_KE_PD = FALSE)
}
\arguments{
\item{g}{Must be an igraph object all vertices must have an attributes called "KE_KED" (Key Event Designator), with values of "MIE", "KE", or "AO"  [igraph object]}

\item{use_KE_PD}{Optional (Default=FALSE): Can use KE_PD ("origin" and "terminus") instead of KE_KED [boolean]}
}
\value{
Edge connectivity of MIE AO pairs [data.frame]
}
\description{
Edge connectivity is the minimum number of edges that must be removed
    in order to disrupt all paths between to vertices. This function computes edge connectivity
    of all MIE/AO pairs
}
\seealso{
Other AOPNet Operations: \code{\link{AOPN.linear.AOPs}},
  \code{\link{topo.layout}}
}
