% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOPN.linear.AOPs.R
\name{AOPN.linear.AOPs}
\alias{AOPN.linear.AOPs}
\title{AOPN.linear.AOPs}
\usage{
AOPN.linear.AOPs(g, use_KE_PD = FALSE)
}
\arguments{
\item{g}{AOP igraph object, with vertex attribute KE_KED (values "MIE", "KE", or "AO"), must also have "names" attribute [igraph object]}

\item{use_KE_PD}{Optional (Default=FALSE): Can use KE_PD ("origin" and "terminus") instead of KE_KED [boolean]}
}
\value{
OUTPUT is a "List of a List" of all paths (by vertex) between all MIE/AO pairs (or origin/terminus pairs)
  List objects are named by the MIE/AO pair involved
  List objects are Lists of all paths (by vertex) between the MIE and AO for which the List is named [list]
}
\description{
Determines all linear AOPs (ie simple paths) between all
possible MIE and AO pairs (or "ORIGIN" and "TERMINUS" if desired)
}
\seealso{
Other AOPNet Operations: \code{\link{AOPN.edge.connectivity}},
  \code{\link{topo.layout}}
}
