% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOPN.short.path.color.R
\name{AOPN.short.path.color}
\alias{AOPN.short.path.color}
\title{AOPN.short.path.color}
\usage{
AOPN.short.path.color(g, fromnode, tonode, loc = F, clr = "purple",
  nonclr = "transparent", weight = NA, all = T)
}
\arguments{
\item{g}{[igraph object]}

\item{fromnode}{The first KE in the path [vertex of igraph object]}

\item{tonode}{The final KE in the path [vertex of igraph object]}

\item{loc}{Leave originial color? TRUE leaves as orginial color, Default=FALSE [boolean]}

\item{clr}{Specified color for path, default="purple" [R color]}

\item{nonclr}{Non-path color, default="transparent" [R Color]}

\item{weight}{Edge weights for shortest path algorithm, default=NA [vector]}

\item{all}{Color all shortest paths or just the first detected? TRUE colors all, default=TRUE [boolean]}
}
\value{
A vector of colors for edges [vector]
}
\description{
Provides colors for all edges in a igraph object, where the shortest path is colored, all other edges are transparent
}
\seealso{
Other Path Operations: \code{\link{edges.from.path}}
}
