% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.scc.R
\name{contract.scc}
\alias{contract.scc}
\title{contract.scc}
\usage{
contract.scc(g)
}
\arguments{
\item{g}{Must be an igraph object all vertices must have an attributes called "KE_KED" (Key Event Designator), with values of "MIE", "KE", or "AO"  [igraph object]}
}
\value{
New condensed igraph object where each strongly connected component has been contracted to a single node [igraph object]
}
\description{
Condense all strongly connected components in a graph
}
