% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.layout.R
\name{topo.layout}
\alias{topo.layout}
\title{topo.layout}
\usage{
topo.layout(g)
}
\arguments{
\item{g}{An igraph object with no strongly connected components [igraph object]}
}
\value{
Layout coordinates for plotting a topologically sorted network (can be used for "layout" argument of the plot function) [array]
}
\description{
Computes a topological ordering layout for visualization of an igraph object, as long as the igraph object is acyclic
}
\seealso{
Other AOPNet Operations: \code{\link{AOPN.edge.connectivity}},
  \code{\link{AOPN.linear.AOPs}}
}
